HawkEdit - moving undo/redo to main menu; undo/redo as a separate class; integrated and working with script edits; no save; no spreadsheet edit integration

+++ codebase/apps/radar/src/HawkEdit/UndoRedoController.cc
@@ -22,7 +22,7 @@
 // ** WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.    
 // *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=* 
 /////////////////////////////////////////////////////////////
-// UndoRedoModel.hh
+// UndoRedoController.hh
 //
 // Coordinates undo and redo activities
 //
@@ -36,239 +36,61 @@
 // 
 ///////////////////////////////////////////////////////////////
 
+ #include <iostream>
 
-#include <iostream>
-#include <sstream>
+#include "UndoRedoController.hh"
 
-#include "UndoRedoModel.hh"
+UndoRedoController::UndoRedoController() {
 
-const string UndoRedoModel::_tmpDir = ".tmp_hawkedit";
-const string UndoRedoModel::_fileDir = "file_";
-
-UndoRedoModel::UndoRedoModel() {
-
-
+	_model = new UndoRedoModel();
+  //_tempDirIndex = -1;
 
 }
 
-UndoRedoModel::~UndoRedoModel() {
-  cerr << "UndoRedoModel destructor called" << endl;
+UndoRedoController::~UndoRedoController() {
+  cerr << "UndoRedoController destructor called" << endl;
 
-  _removeTempDirs();
-}
-
-
-void UndoRedoModel::_createDirectory(string path) {
-    RadxPath tmpDir;
-    tmpDir.setDirectory(path);
-    int result = tmpDir.makeDirRecurse();
-    if (result != 0) {
-      throw std::invalid_argument("cannot make temporary directory for undo and redo");
-    }
-}
-
-void UndoRedoModel::setBaseDir(string path, int nFiles) {
-  baseDir.clear();
-  baseDir.setDirectory(path);
-  currentVersion.clear();
-  currentVersion.resize(nFiles);
-
-  for (int i=0; i<nFiles; i++) {
-    currentVersion.at(i) = 0;
-    stringstream ss;
-    ss << path << "/" << _tmpDir << "/" << _fileDir << i;
-    _createDirectory(ss.str());
-//    RadxPath tmpDir;
-//    tmpDir.setDirectory(ss.str());
-//    int result = tmpDir.makeDirRecurse();
-//    if (result != 0) {
-//      throw std::invalid_argument("cannot make temporary directory for undo and redo");
-//    }
-  }
-  _currentBatchIndex = -1;
-  //makeNewBatch(); // record the intial watermark
-}
-
-
-void UndoRedoModel::clear() {
-  // remove .tmp dir // removes all temp versions; clears the stacks
-  _removeTempDirs();
-  currentVersion.clear();
-  batches.clear();
-}
-
-string UndoRedoModel::_constructFullTempPath(int fileNum, int version) {
-  stringstream ss;
-  ss << baseDir.getDirectory() << "/" << _tmpDir << "/" << _fileDir << fileNum << 
-    "/v" << version;
-  return ss.str();
-}
-
-// version 0 is the original file
-
-string UndoRedoModel::undo(int fileNum) {
-  //find the previous version of the file in the temp dir; change to this file.
-  int previousN = currentVersion.at(fileNum) - 1;
-  if (previousN <= 0) {
-    currentVersion.at(fileNum) = 0;
-    return ""; // indicate we are at the end of versions
-  } else {
-    currentVersion.at(fileNum) = previousN;
-    return _constructFullTempPath(fileNum, previousN);
+  if (_model != NULL) { 
+    delete _model;
   }
 }
 
-string UndoRedoModel::redo(int fileNum) {
-  //find the next version of the file in the temp dir; change to this file.
-  int nextN = currentVersion.at(fileNum) + 1;
-  string nextFullPath = _constructFullTempPath(fileNum, nextN);
-  RadxPath nextPath;
-  //nextPath.setDirectory(nextFullPath);
-  if (!nextPath.exists(nextFullPath)) {
-    return ""; // indicate we are at the end of versions
-  } else {
-    currentVersion.at(fileNum) = nextN;
-    return nextFullPath;
-  }
-  //writeToVersion(currentFile, tempFile_N)
-  //set currentVersion for the file.
+void UndoRedoController::reset(string path, int nFiles) {
+  _model->setBaseDir(path, nFiles);
 }
 
-//void UndoRedoModel::moveToNextVersion ???
-
-
-string UndoRedoModel::moveToNextVersion(int fileNum) { //  or (path/file)
-  //.tmp/file_1/vN
-  currentVersion.at(fileNum) += 1;
-  int nextN = currentVersion.at(fileNum); 
-  string path = _constructFullTempPath(fileNum, nextN);
-  _createDirectory(path);
-  return path;
+string UndoRedoController::getPreviousVersion(int fileIndex) {
+  return _model->undo(fileIndex);
 }
 
-bool UndoRedoModel::_validFileNum(int fileNum) {
-  return ((fileNum >= 0) && (fileNum < currentVersion.size()));
-} 
-
-string UndoRedoModel::getCurrentVersion(int fileNum) {
-  if (_validFileNum(fileNum)) {
-    int version = currentVersion.at(fileNum);
-    if (version > 0) {
-      return _constructFullTempPath(fileNum, version);
-    } else {
-      return "";  // version 0 indicates the original file
-    }
-  } else {
-    return "";
-  }
+string UndoRedoController::getNextVersion(int fileIndex) {
+  return _model->redo(fileIndex);
 }
 
-int UndoRedoModel::getCurrentVersionNum(int fileNum) {
-  if (_validFileNum(fileNum)) {
-    return currentVersion.at(fileNum);
-  } else {
-    return -1;
-  }
+string UndoRedoController::getCurrentVersion(int fileIndex) {
+  return _model->getCurrentVersion(fileIndex);
 }
-  //save(int fileNum)
-  //fn = getCurrentVersion(file#) 
-  // move (fn, savePath)
 
-  //saveAll(bool overwrite, path)
-  //saves current version of all files
+string UndoRedoController::getNewVersion(int fileIndex) {
+  return _model->moveToNextVersion(fileIndex);
+}
+
+void UndoRedoController::waterMarkVersion() {
+  _model->makeNewBatch();
+}
 
 /*
-void UndoRedoModel::fetchArchiveFiles(string seedPath, string seedFileName,
-    string fullUrl, bool keepTimeRange) {
-  
-  int nFilesFound = _model->findArchiveFileList(seedPath, keepTimeRange);
-
-  if (nFilesFound < 1) {
-    _model->recoverFromNoDatDirFormat(fullUrl);
-  }
-
-  _setGuiFromArchiveStartTime();  
-  _setGuiFromArchiveEndTime();
-  _model->setSelectedFile(seedFileName);
-  _setGuiFromSelectedTime();
-
-  _view->setNTicks(_model->getNArchiveFiles());
-  //setSliderPosition(_model->getPositionOfSelection());
-
-  _view->showTimeControl();
-}
-
-*/
-
-// move all the version numbers from the batch to the currentVersion 
-void UndoRedoModel::batchUndo() {
-
-  if (_currentBatchIndex > 0) {
-    _currentBatchIndex -= 1;
-    vector<int> *previousBatch = batches.at(_currentBatchIndex);
-    for (int i=0; i<currentVersion.size(); i++) {
-      currentVersion.at(i) = previousBatch->at(i);
-    }
-  } else {
-    throw std::invalid_argument("no more batch undo");
-  }
-}
-
-void UndoRedoModel::batchRedo() {
-  if (_currentBatchIndex < batches.size() -1) {
-    _currentBatchIndex += 1;
-    vector<int> *previousBatch = batches.at(_currentBatchIndex);
-    for (int i=0; i<currentVersion.size(); i++) {
-      currentVersion.at(i) = previousBatch->at(i);
-    }
-  } else {
-    throw std::invalid_argument("no more batch undo");
-  }
-}
-
-// add another batch to the batches store;
-// take the current version of every file, and increment by one.
-// these are the versions to use when writing batch edited files.
-void UndoRedoModel::makeNewBatch() {
-  size_t nFiles = currentVersion.size();
-  vector<int> *newBatch = new vector<int>;
-  newBatch->resize(nFiles);
-  // push previous versions onto stack
-  // these will always be -1 from the current versions
-  for (int i=0; i<nFiles; i++) {
-    newBatch->at(i) = currentVersion.at(i) -1;
-  }
-  batches.push_back(newBatch);
-  _currentBatchIndex += 1;
-
-  // push the current versions onto the stack
-  newBatch = new vector<int>;
-  newBatch->resize(nFiles);
-  for (int i=0; i<nFiles; i++) {
-    newBatch->at(i) = currentVersion.at(i);
-  }
-  batches.push_back(newBatch);
-  _currentBatchIndex += 1;
-}
-
-// get the  version for this file
-// can save to the file path returned
-//string batchGetVersion(int fileNum) {
-//  return 
-//}
-
-/*
-string &UndoRedoModel::getSelectedArchiveFile() {
+string &UndoRedoController::getSelectedArchiveFile() {
   return _model->getSelectedArchiveFile();
 }
 
-string UndoRedoModel::getSelectedPath() {
+string UndoRedoController::getSelectedPath() {
   return _model->getCurrentPath();
 }
 
 // compare nextTempDir base directory to the base directory
 // in the temp stack
-bool UndoRedoModel::_isDifferentBaseDir(string nextTempDir) {
+bool UndoRedoController::_isDifferentBaseDir(string nextTempDir) {
   bool different = true;
   if (_tempDirStack.size() > 0) {
     // then there is a previous base directory
@@ -312,7 +134,7 @@ bool UndoRedoModel::_isDifferentBaseDir(string nextTempDir) {
 // push base dir on the stack, and clear the stack when
 // setting a new base dir
 
-string UndoRedoModel::getTempDir() {
+string UndoRedoController::getTempDir() {
   string nextTempDir = _model->getTempDir();
 
   if (_tempDirStack.empty()) {
@@ -337,13 +159,13 @@ string UndoRedoModel::getTempDir() {
   return nextTempDir;
 }
 
-string UndoRedoModel::getSelectedArchiveFileName() {
+string UndoRedoController::getSelectedArchiveFileName() {
   return _model->getSelectedArchiveFileName();
 }
 
 // return true if the current directory is a temp dir
 // i.e. of the form .../.tmp_N/yyyymmdd
-bool UndoRedoModel::isSelectedFileInTempDir() {
+bool UndoRedoController::isSelectedFileInTempDir() {
   return (_tempDirIndex >= 0);
 }
 
@@ -359,7 +181,7 @@ bool UndoRedoModel::isSelectedFileInTempDir() {
 // i.e. idx < 0; or use a stack????
 // return empty string, if there are no more directories
 // in the stack
-string UndoRedoModel::getPreviousTempDir() {
+string UndoRedoController::getPreviousTempDir() {
   if (_tempDirIndex <= 0) {
     return "";
   } else {
@@ -374,7 +196,7 @@ string UndoRedoModel::getPreviousTempDir() {
 // i.e. idx < 0; or use a stack????
 // return empty string, if there are no more directories
 // in the stack
-string UndoRedoModel::getNextTempDir() {
+string UndoRedoController::getNextTempDir() {
   if (_tempDirIndex >= _tempDirStack.size()-1) {
     return "";
   } else {
@@ -384,7 +206,7 @@ string UndoRedoModel::getNextTempDir() {
   }
 }
 
-void UndoRedoModel::_resetTempStack() {
+void UndoRedoController::_resetTempStack() {
 
   // delete all temporary directories in stack
   _removeTempDirs();
@@ -392,17 +214,17 @@ void UndoRedoModel::_resetTempStack() {
   _tempDirStack.clear();
 }
 
-void UndoRedoModel::_setBaseDirTempStack() {
+void UndoRedoController::_setBaseDirTempStack() {
   // set the base directory for any edits
   _tempDirIndex = 0;
   _tempDirStack.push_back(_model->getCurrentPath());
 }
 
-bool UndoRedoModel::moreFiles() {
+bool UndoRedoController::moreFiles() {
   return _model->moreFiles();
 }
 
-void UndoRedoModel::getBounds(bool useTimeRange, int *firstArchiveFileIndex,
+void UndoRedoController::getBounds(bool useTimeRange, int *firstArchiveFileIndex,
     int *lastArchiveFileIndex) {
   if (useTimeRange) {
     *firstArchiveFileIndex = 0;
@@ -414,7 +236,7 @@ void UndoRedoModel::getBounds(bool useTimeRange, int *firstArchiveFileIndex,
 }
 
 
-bool UndoRedoModel::_isTempDir(string *path) {
+bool UndoRedoController::_isTempDir(string *path) {
   bool found = false;
   if (path->find(".tmp_") != string::npos) {
     found = true;
@@ -424,28 +246,23 @@ bool UndoRedoModel::_isTempDir(string *path) {
 
 // move/rename the current selected temp directory
 // to the newName
-void UndoRedoModel::replaceSelectedTempPath(string newName) {
+void UndoRedoController::replaceSelectedTempPath(string newName) {
   _tempDirStack[_tempDirIndex] = newName;
 }
-*/
 
 // before changing to a new base directory, prompt to save any temp dirs?
 // otherwise, the temp dirs will be deleted.
 // new method ... movingToNewBaseDir ...
 // if moving to new base dir, clear the stack
-void UndoRedoModel::_removeTempDirs() {
-
+void UndoRedoController::_removeTempDirs() {
   //vector<string>::iterator it;
-  //if (_tempDirStack.size() > 1) {
+  if (_tempDirStack.size() > 1) {
 
     // clean up temp dirs
     string command = "rm -r ";
-    command.append(baseDir.getDirectory()); //  getSelectedPath());
-    command.append("/");
-    command.append(_tmpDir);
+    command.append(_tempDirStack.at(0)); //  getSelectedPath());
+    command.append("/.tmp_*");
 
-    //cout << command << endl;
-    //int result = 0;
     int result = std::system(command.c_str()); // execute the UNIX command "mv -f oldName newName"
 
     //  The rm utility exits 0 on success, and >0 if an error occurs.
@@ -458,7 +275,9 @@ void UndoRedoModel::_removeTempDirs() {
     } else {
 
     }
-  //}
+  }
   //
 }
+
+*/
